fix(frontend): add better commit message, refactor userEvent code and change content and name of setting.json file
The purpose of the first test is to verify that the component is rendered correctly by examining the contents of the component. I want to achieve this by checking the correct content of the heading, paragraph and button tags.The purpose of the second test is to check whether the goBack function will be called after clicking on the go-back button. I changed the code in this test to invoke userEvent.setup() before rendering the component which is recommended way according to the https://testing-library.com/docs/user-event/intro/ documentation. I renamed the file from setting.json to settings.json, it was just a spelling error. I changed the content of this file, which is now identical to the Orbison application.